refactored openapi
