+++ docs/usage/1-route-handlers.md
@@ -43,7 +43,7 @@ The `route` decorator receives the following kwargs -
   response. Defaults to `MediaType.JSON`.
 * `response_class`: The response class to use. The value must be either a `Starlette` response class or a class that
   extends it. Defaults to `starlite.Response`.
-* `response_headers`: A _dataclass_, _TypedDict_ or _pydantic model_ that describes the response headers. This value is
+* `response_headers`: A _dataclass_ or _pydantic model_ that describes the response headers. This value is
   used only in the OpenAPI schema. Defaults to `None`.
 * `dependencies`: A dictionary mapping dependency injection Providers to keys. Defaults to `None`.
 * `include_in_schema`: A boolean flag dictating whether the given route handler will appear in the generated OpenAPI
